auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1